requirements.txt
websockets==12.0
redis==5.0.1
pandas==2.2.0
numpy==1.26.3
xgboost==2.0.3
scikit-learn==1.4.0
transformers==4.38.0
python-dotenv==1.0.1
requests==2.31.0
ta-lib-binary==0.4.33
pytest==8.0.0